% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{cleanAPIData}
\alias{cleanAPIData}
\title{Unnest JSON pulled from Horizon API.}
\usage{
cleanAPIData(API_data)
}
\arguments{
\item{API_data}{JSON output from the Horizon API.}
}
\value{
Formatted data frame.
}
\description{
Unnest JSON pulled from Horizon API.
}
