% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getAssets}
\alias{getAssets}
\title{Get assets information from the Horizon API}
\usage{
getAssets(assetCode = NULL, assetIssuer = NULL, limit = NULL,
  order = NULL)
}
\arguments{
\item{assetCode}{Code for an asset available on the API}

\item{limit}{The number of elements that will be in the resulting data frame.}

\item{order}{parameter to pass to the API call can be NULL, "asc", or "desc". If NULL the API default is "asc"}

\item{assetCode}{Issuer for an asset available on the API}
}
\value{
Formatted data frame
}
\description{
Get assets information from the Horizon API
}
