% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getLedgerData}
\alias{getLedgerData}
\title{Get historical ledger data for a specified ledger}
\usage{
getLedgerData(ledger, order = NULL, limit = NULL)
}
\arguments{
\item{ledger}{Ledger ID}

\item{order}{parameter to pass to the API call can be NULL, "asc", or "desc". If NULL the API default is "asc"}

\item{limit}{The number of elements that will be in the resulting data frame.}
}
\value{
List containing information, transactions, operations, payments, and effects for the ledger.
}
\description{
Get historical ledger data for a specified ledger
}
