% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getLineage}
\alias{getLineage}
\title{Get account creation history for a specified account}
\usage{
getLineage(stellarAddress, generationsMax = 10, generation = NULL,
  lineage = NULL)
}
\arguments{
\item{stellarAddress}{Stellar address}

\item{generationsMax}{The maximum number of generations to return}

\item{generation}{Indicator informing the number of generations for recurring calls}

\item{lineage}{output from previous getLineage() function to append new data to}
}
\value{
Formatted data frame giving parent account, child account and creation date
}
\description{
Get account creation history for a specified account
}
