% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getOperationData}
\alias{getOperationData}
\title{Get account creation history for a specified account}
\usage{
getOperationData(operation, order = NULL, limit = NULL)
}
\arguments{
\item{order}{parameter to pass to the API call can be NULL, "asc", or "desc". If NULL the API default is "asc"}

\item{limit}{The number of elements that will be in the resulting data frame.}

\item{address}{Stellar address}
}
\value{
List containing information, transactions, operations, payments, effects, and offers for the account.
}
\description{
Get account creation history for a specified account
}
