% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getOperations}
\alias{getOperations}
\title{Get historical operations from the Stellar API}
\usage{
getOperations(order = NULL, limit = NULL)
}
\arguments{
\item{order}{parameter to pass to the API call can be NULL, "asc", or "desc". If NULL the API default is "asc"}

\item{limit}{The number of elements that will be in the resulting data frame.}
}
\value{
Formatted data frame.
}
\description{
Get historical operations from the Stellar API
}
