% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getStellarData}
\alias{getStellarData}
\title{A recursive function that gets the last N entries from a Horizon API endpoint}
\usage{
getStellarData(endpoint = NULL, limit = NULL, link = NULL, data = NULL,
  order = "desc")
}
\arguments{
\item{endpoint}{Endpoint from the Horizon API such as 'ledgers' or 'transactions'.}

\item{limit}{The number of elements that will be in the resulting data frame.}

\item{link}{Direct link to the Horizon API such as 'https://horizon.stellar.org/ledgers'}

\item{data}{Dataframe from getStellarData() to append new data to}

\item{order}{Either 'asc' or 'desc' which is passed as a parameter to the API call.}
}
\value{
Formatted data frame.
}
\description{
A recursive function that gets the last N entries from a Horizon API endpoint
}
\examples{
ledger_data <- getStellarData('ledgers',limit=500)
effects_data <- getStellarData('effects',limit=250)
transaction_effects_data <- getStellarData(link="https://horizon.stellar.org/transactions/f93dfc81770019f2e842c6a1f3ecb2746db8e2b9b5b30b316128e6504de52f06/effects")
}
