% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getTrades}
\alias{getTrades}
\title{Get historical trades from the Stellar API}
\usage{
getTrades(order = NULL, limit = NULL, offerid = NULL)
}
\arguments{
\item{order}{Parameter to pass to the API call can be NULL, "asc", or "desc". If NULL the API default is "asc"}

\item{limit}{The number of elements that will be in the resulting data frame.}

\item{offerid}{Offer id for a trade}
}
\value{
Formatted data frame.
}
\description{
Get historical trades from the Stellar API
}
