% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getTransactionData}
\alias{getTransactionData}
\title{Get transaction data for a specified transaction}
\usage{
getTransactionData(transaction, order = NULL, limit = NULL)
}
\arguments{
\item{transaction}{Transaction ID}

\item{order}{parameter to pass to the API call can be NULL, "asc", or "desc". If NULL the API default is "asc"}

\item{limit}{The number of elements that will be in the resulting data frame.}
}
\value{
List containing information, operations, and effects for the transaction.
}
\description{
Get transaction data for a specified transaction
}
