% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{unnestDF}
\alias{unnestDF}
\title{Unnest JSON pulled from Horizon API.}
\usage{
unnestDF(data)
}
\arguments{
\item{data}{Embedded records data from Horizon API.}
}
\value{
Formatted data frame.
}
\description{
Unnest JSON pulled from Horizon API.
}
